% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{is_absolute_path}
\alias{is_absolute_path}
\title{Check whether a file path is absolute}
\usage{
is_absolute_path(path)
}
\arguments{
\item{path}{Character. The path to check.}
}
\value{
Logical. \code{TRUE} if absolute, \code{FALSE} if relative.
}
\description{
Check whether a file path is absolute
}
