% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{col_md}
\alias{col_md}
\alias{col_md2}
\alias{col_md3}
\alias{col_md4}
\alias{col_bulk}
\alias{col_diffuse}
\alias{col_good}
\alias{col_warn}
\alias{col_bad}
\alias{col_group}
\alias{ext_actorPowerDrivers}
\alias{ext_instQualityDrivers}
\alias{ext_vdemDrivers}
\alias{ext_instQualityDrivers_all}
\alias{ext_controlDrivers}
\title{PFM report palette and driver-name constants}
\format{
Character scalars / vectors.
}
\usage{
col_md

col_md2

col_md3

col_md4

col_bulk

col_diffuse

col_good

col_warn

col_bad

col_group

ext_actorPowerDrivers

ext_instQualityDrivers

ext_vdemDrivers

ext_instQualityDrivers_all

ext_controlDrivers
}
\description{
Exported colour and driver-name vectors used across the report templates.
}
